refactor(parser): add more tests, reduce char grabber size
